Backend/requirement.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
chromadb
pypdf
tiktoken
python-dotenv
langchain-community
langchain-chroma
sentence-transformers